fix: move root to 'views'
